* t_an_to_ln.c: Include <stdio.h> for prototypes
	* locate_kdc.c (krb5_locate_srv_dns): Remove unused variable.

	* changepw.c, sendto_kdc.c, sn2princ.c: Add parentheses around
	assignment used as truth value.

	* ccdefname.c (get_from_os): Cast return from getuid() to long and
	indicate use of format in sprintf.

gcc -Wall complaints.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12482 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/ccdefname.c
@@ -224,7 +224,7 @@ cleanup:
 #if !(defined(_MSDOS) || defined(_WIN32))
 static krb5_error_code get_from_os(char *name_buf, int name_size)
 {
-	sprintf(name_buf, "FILE:/tmp/krb5cc_%d", getuid());
+	sprintf(name_buf, "FILE:/tmp/krb5cc_%ld", (long) getuid());
 	return 0;
 }
 #endif
